Update Submission_guidelines.md (#146)
* Update Submission_guidelines.md

Separate the Training metrics from the Checkpoint metrics.  Also, define the checkpoint metrics as B/W during writes and reads, but include the switchover time that some storage systems may require between writing a checkpoint and reading it from another host name.

Linked to issue #129

* Update Submission_guidelines.md

Added some clarifying text that the remapping time will be added to the read/recover benchmark metric.

* Update Submission_guidelines.md

Be clear that this only applies to Training so there is no confusion on how this applies to Checkpointing.

* Update Submission_guidelines.md

Remove a change that belongs in a different PR